introduced new deployment result type and moved logic out of reconciler
